+++ app/move/views/view.njk
@@ -12,7 +12,7 @@
   {{ govukBackLink({
     text: t("actions:back_to_dashboard"),
     classes: "app-print--hide",
-    href: "/"
+    href: "/moves"
   }) }}
 
   <a href="javascript:window.print()" class="app-!-float-right app-print--hide govuk-!-margin-top-3">
