+++ app/moves/index.js
@@ -4,7 +4,12 @@ const router = require('express').Router()
 // Local dependencies
 const { protectRoute } = require('../../common/middleware/permissions')
 const { download, list } = require('./controllers')
-const { setMoveDate, setFromLocation, setMovesByDate } = require('./middleware')
+const {
+  redirectUsers,
+  setMoveDate,
+  setFromLocation,
+  setMovesByDate,
+} = require('./middleware')
 
 const uuidRegex =
   '[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}'
@@ -14,7 +19,13 @@ router.param('locationId', setFromLocation)
 
 // Define routes
 router.use(setMoveDate)
-router.get('/', protectRoute('moves:view:all'), setMovesByDate, list)
+router.get(
+  '/',
+  redirectUsers,
+  protectRoute('moves:view:all'),
+  setMovesByDate,
+  list
+)
 router.get(
   `/:locationId(${uuidRegex})`,
   protectRoute('moves:view:by_location'),
